+++ resources/js/Components/Global/Search/DefaultSearchResult.vue
@@ -1,0 +1,7 @@
+<template>
+<!-- This is a placeholder. -->
+  <div>Default Search Result</div>
+</template>
+<script setup>
+
+</script>